function return the floor of the float

+++ python_variable_annotations/2-floor.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
-import math
 
 """  function returns the floor of the float """
+import math
 
 
 def floor(n: float) -> int:
     """ Return the floor of a float """
-    return int(math.floor(n))
+    return math.floor(n)
